fix: refactor code structure and types
