+++ app/shared/components/configurable-list/item-wrapper/ItemWrapper.styles.ts
